+++ ui/CHANGELOG.md
@@ -26,6 +26,7 @@ ___
 
 ### 🚀 Added
 
+- PostHog analytics integration [(#8357)](https://github.com/prowler-cloud/prowler/pull/8357)
 - Mutelist configuration form [(#8190)](https://github.com/prowler-cloud/prowler/pull/8190)
 - SAML login integration [(#8203)](https://github.com/prowler-cloud/prowler/pull/8203)
 - Resource view [(#7760)](https://github.com/prowler-cloud/prowler/pull/7760)